Move select verse/version to their own components.
Create SelectVerse and SelectVersion components to consolidate the
functionality.

Update lib/store to provide 3 stores for the version in anticipation of
multi-column, parallel version presentations.

Update Version to make use of the new components and store provisions.
